# Reduce time to allocate/destroy a large chunk