--- a/Makefile
+++ b/Makefile
@@ -113,6 +113,7 @@ endif
 PONY_BUILD_DIR   ?= build/$(config)
 PONY_SOURCE_DIR  ?= src
 PONY_TEST_DIR ?= test
+PONY_BENCHMARK_DIR ?= benchmark
 
 ifdef use
   ifneq (,$(filter $(use), valgrind))
@@ -237,6 +238,7 @@ $(shell mkdir -p $(PONY_BUILD_DIR))
 lib   := $(PONY_BUILD_DIR)
 bin   := $(PONY_BUILD_DIR)
 tests := $(PONY_BUILD_DIR)
+benchmarks := $(PONY_BUILD_DIR)
 obj   := $(PONY_BUILD_DIR)/obj
 
 # Libraries. Defined as
@@ -294,13 +296,16 @@ endif
 libgtest := $(lib)
 libgtest.dir := lib/gtest
 libgtest.files := $(libgtest.dir)/gtest_main.cc $(libgtest.dir)/gtest-all.cc
+libgbenchmark := $(lib)
+libgbenchmark.dir := lib/gbenchmark
+libgbenchmark.files := $(libgbenchmark.dir)/gbenchmark_main.cc $(libgbenchmark.dir)/gbenchmark-all.cc
 
 # We don't add libponyrt here. It's a special case because it can be compiled
 # to LLVM bitcode.
 ifeq ($(OSTYPE), linux)
-  libraries := libponyc libponyrt-pic libgtest
+  libraries := libponyc libponyrt-pic libgtest libgbenchmark
 else
-  libraries := libponyc libgtest
+  libraries := libponyc libgtest libgbenchmark
 endif
 
 # Third party, but prebuilt. Prebuilt libraries are defined as
@@ -340,6 +345,12 @@ libponyrt.tests := $(tests)
 
 tests := libponyc.tests libponyrt.tests
 
+# Benchmark suites are directly attached to the libraries they test.
+libponyc.benchmarks  := $(benchmarks)
+libponyrt.benchmarks := $(benchmarks)
+
+benchmarks := libponyc.benchmarks libponyrt.benchmarks
+
 # Define include paths for targets if necessary. Note that these include paths
 # will automatically apply to the test suite of a target as well.
 libponyc.include := -I src/common/ -I src/libponyrt/ $(llvm.include)
@@ -351,8 +362,14 @@ libponyc.tests.include := -I src/common/ -I src/libponyc/ $(llvm.include) \
   -isystem lib/gtest/
 libponyrt.tests.include := -I src/common/ -I src/libponyrt/ -isystem lib/gtest/
 
+libponyc.benchmarks.include := -I src/common/ -I src/libponyc/ \
+  $(llvm.include) -isystem lib/gbenchmark/include/
+libponyrt.benchmarks.include := -I src/common/ -I src/libponyrt/ -isystem \
+  lib/gbenchmark/include/
+
 ponyc.include := -I src/common/ -I src/libponyrt/ $(llvm.include)
 libgtest.include := -isystem lib/gtest/
+libgbenchmark.include := -isystem lib/gbenchmark/include/
 
 ifneq (,$(filter $(OSTYPE), osx freebsd))
   libponyrt.include += -I /usr/local/include
@@ -367,6 +384,12 @@ libponyc.tests.buildoptions = -D__STDC_CONSTANT_MACROS
 libponyc.tests.buildoptions += -D__STDC_FORMAT_MACROS
 libponyc.tests.buildoptions += -D__STDC_LIMIT_MACROS
 
+libponyc.benchmarks.buildoptions = -D__STDC_CONSTANT_MACROS
+libponyc.benchmarks.buildoptions += -D__STDC_FORMAT_MACROS
+libponyc.benchmarks.buildoptions += -D__STDC_LIMIT_MACROS
+
+libgbenchmark.buildoptions := -DHAVE_POSIX_REGEX
+
 ponyc.buildoptions = $(libponyc.buildoptions)
 
 ifeq ($(OSTYPE), linux)
@@ -375,16 +398,21 @@ endif
 
 # target specific disabling of build options
 libgtest.disable = -Wconversion -Wno-sign-conversion -Wextra
+libgbenchmark.disable = -Wconversion -Wno-sign-conversion -Wextra
 
 # Link relationships.
 ponyc.links = libponyc libponyrt llvm
 libponyc.tests.links = libgtest libponyc libponyrt llvm
 libponyrt.tests.links = libgtest libponyrt
+libponyc.benchmarks.links = libgbenchmark libponyc libponyrt llvm
+libponyrt.benchmarks.links = libgbenchmark libponyrt
 
 ifeq ($(OSTYPE),linux)
   ponyc.links += pthread dl
   libponyc.tests.links += pthread dl
   libponyrt.tests.links += pthread dl
+  libponyc.benchmarks.links += pthread dl
+  libponyrt.benchmarks.links += pthread dl
 endif
 
 ifneq (, $(DTRACE))
@@ -393,9 +421,11 @@ endif
 
 # Overwrite the default linker for a target.
 ponyc.linker = $(CXX) #compile as C but link as CPP (llvm)
+libponyc.benchmarks.linker = $(CXX)
+libponyrt.benchmarks.linker = $(CXX)
 
 # make targets
-targets := $(libraries) libponyrt $(binaries) $(tests)
+targets := $(libraries) libponyrt $(binaries) $(tests) $(benchmarks)
 
 .PHONY: all $(targets) install uninstall clean stats deploy prerelease
 all: $(targets)
@@ -405,6 +435,8 @@ all: $(targets)
 libponyc.depends := libponyrt
 libponyc.tests.depends := libponyc libgtest
 libponyrt.tests.depends := libponyrt libgtest
+libponyc.benchmarks.depends := libponyc libgbenchmark
+libponyrt.benchmarks.depends := libponyrt libgbenchmark
 ponyc.depends := libponyc libponyrt
 
 # Generic make section, edit with care.
@@ -435,6 +467,9 @@ define DIRECTORY
   else ifneq ($$(filter $(1),$(tests)),)
     sourcedir := $(PONY_TEST_DIR)/$(subst .tests,,$(1))
     outdir := $(obj)/tests/$(subst .tests,,$(1))
+  else ifneq ($$(filter $(1),$(benchmarks)),)
+    sourcedir := $(PONY_BENCHMARK_DIR)/$(subst .benchmarks,,$(1))
+    outdir := $(obj)/benchmarks/$(subst .benchmarks,,$(1))
   else
     sourcedir := $(PONY_SOURCE_DIR)/$(1)
   endif
@@ -644,6 +679,20 @@ endef
 
 $(eval $(call EXPAND_UNINSTALL))
 
+ifdef verbose
+  bench_verbose = -DCMAKE_VERBOSE_MAKEFILE=true
+endif
+
+ifeq ($(lto),yes)
+  bench_lto = -DBENCHMARK_ENABLE_LTO=true
+endif
+
+benchmark: all
+	@echo "Running libponyc benchmarks..."
+	@$(PONY_BUILD_DIR)/libponyc.benchmarks
+	@echo "Running libponyrt benchmarks..."
+	@$(PONY_BUILD_DIR)/libponyrt.benchmarks
+
 test: all
 	@$(PONY_BUILD_DIR)/libponyc.tests
 	@$(PONY_BUILD_DIR)/libponyrt.tests
@@ -756,17 +805,20 @@ help:
 	@echo '   dtrace'
 	@echo
 	@echo 'TARGETS:'
-	@echo '  libponyc          Pony compiler library'
-	@echo '  libponyrt         Pony runtime'
-	@echo '  libponyrt-pic     Pony runtime -fpic'
-	@echo '  libponyc.tests    Test suite for libponyc'
-	@echo '  libponyrt.tests   Test suite for libponyrt'
-	@echo '  ponyc             Pony compiler executable'
+	@echo '  libponyc               Pony compiler library'
+	@echo '  libponyrt              Pony runtime'
+	@echo '  libponyrt-pic          Pony runtime -fpic'
+	@echo '  libponyc.tests         Test suite for libponyc'
+	@echo '  libponyrt.tests        Test suite for libponyrt'
+	@echo '  libponyc.benchmarks    Benchmark suite for libponyc'
+	@echo '  libponyrt.benchmarks   Benchmark suite for libponyrt'
+	@echo '  ponyc                  Pony compiler executable'
 	@echo
-	@echo '  all               Build all of the above (default)'
-	@echo '  test              Run test suite'
-	@echo '  install           Install ponyc'
-	@echo '  uninstall         Remove all versions of ponyc'
-	@echo '  stats             Print Pony cloc statistics'
-	@echo '  clean             Delete all build files'
+	@echo '  all                    Build all of the above (default)'
+	@echo '  test                   Run test suite'
+	@echo '  benchmark              Build and run benchmark suite'
+	@echo '  install                Install ponyc'
+	@echo '  uninstall              Remove all versions of ponyc'
+	@echo '  stats                  Print Pony cloc statistics'
+	@echo '  clean                  Delete all build files'
 	@echo