--- a/lib/gbenchmark/gbenchmark_main.cc
+++ b/lib/gbenchmark/gbenchmark_main.cc
@@ -1,0 +1,5 @@
+#include <stdio.h>
+
+#include "benchmark/benchmark.h"
+
+BENCHMARK_MAIN()