* Enhance large chunk pagemap setting by setting multiple pagemaps
  in bulk to avoid unnecessary work and reduce time to allocate
  and destroy a large chunk.
* Enhance heap tests to ensure pagemaps are set correctly for
  large chunks.

NOTE: This PR builds on top of PR #1565 that added benchmarks and includes that commit (which should become a no-op if this is applied after that one is merged). The changes for this PR (excluding the commit adding the benchmarks) is only to 4 files (see: https://github.com/dipinhora/ponyc/commit/319304248a40caeb384de5bfc053a0f8d491a2dc).

Old performance:

```
2: Test command: /home/vagrant/dhponyc/build/release/bench/bench/heap "--benchmark_min_time=0.05"
2: Test timeout computed to be: 9.99988e+06
2: Run on (2 X 2194.92 MHz CPU s)
2: 2017-02-07 20:57:57
2: Benchmark                            Time           CPU Iterations
2: ------------------------------------------------------------------
2: HeapBench/HeapAlloc/32             453 ns        459 ns     148689   2.07673M items/s
2: HeapBench/HeapAlloc/64             459 ns        461 ns     151298   2.06945M items/s
2: HeapBench/HeapAlloc/128            457 ns        455 ns     151817   2.09566M items/s
2: HeapBench/HeapAlloc/256            463 ns        463 ns     153339   2.06053M items/s
2: HeapBench/HeapAlloc/512            454 ns        454 ns     145397    2.1024M items/s
2: HeapBench/HeapAlloc/1024           457 ns        453 ns     152911   2.10356M items/s
2: HeapBench/HeapAlloc/2k             468 ns        471 ns     143924   2.02645M items/s
2: HeapBench/HeapAlloc/4k             484 ns        488 ns     141902   1.95226M items/s
2: HeapBench/HeapAlloc/8k             534 ns        539 ns     136094   1.76998M items/s
2: HeapBench/HeapAlloc/16k            576 ns        580 ns     122477   1.64341M items/s
2: HeapBench/HeapAlloc/32k            677 ns        681 ns     105084   1.40113M items/s
2: HeapBench/HeapAlloc/64k            892 ns        895 ns      80432   1091.67k items/s
2: HeapBench/HeapAlloc/128k          1272 ns       1274 ns      55342     766.4k items/s
2: HeapBench/HeapAlloc/256k          2070 ns       2067 ns      28350   472.437k items/s
2: HeapBench/HeapAlloc/512k          3754 ns       3746 ns      17214   260.671k items/s
2: HeapBench/HeapAlloc/1024k         6787 ns       6790 ns      10278   143.826k items/s
2: HeapBench/HeapDestroy/32           457 ns        451 ns     153422   2.11238M items/s
2: HeapBench/HeapDestroy/64           451 ns        447 ns     152297   2.13324M items/s
2: HeapBench/HeapDestroy/128          452 ns        448 ns     152840   2.12661M items/s
2: HeapBench/HeapDestroy/256          454 ns        450 ns     150401   2.11714M items/s
2: HeapBench/HeapDestroy/512          452 ns        449 ns     147434   2.12523M items/s
2: HeapBench/HeapDestroy/1024         456 ns        452 ns     151835    2.1098M items/s
2: HeapBench/HeapDestroy/2k           469 ns        471 ns     149137   2.02491M items/s
2: HeapBench/HeapDestroy/4k           480 ns        473 ns     141753   2.01449M items/s
2: HeapBench/HeapDestroy/8k           511 ns        508 ns     139446    1.8776M items/s
2: HeapBench/HeapDestroy/16k          562 ns        557 ns     114821   1.71358M items/s
2: HeapBench/HeapDestroy/32k          683 ns        679 ns     106455   1.40467M items/s
2: HeapBench/HeapDestroy/64k          871 ns        866 ns      75411   1.10168M items/s
2: HeapBench/HeapDestroy/128k        1308 ns       1304 ns      56975   748.999k items/s
2: HeapBench/HeapDestroy/256k        2086 ns       2084 ns      32303   468.497k items/s
2: HeapBench/HeapDestroy/512k        4040 ns       4034 ns      19153   242.096k items/s
2: HeapBench/HeapDestroy/1024k       7504 ns       7499 ns       8504   130.226k items/s
```


New performance:
```
2: Test command: /home/vagrant/dhponyc/build/release/bench/bench/heap "--benchmark_min_time=0.05"
2: Test timeout computed to be: 9.99988e+06
2: Run on (2 X 2194.92 MHz CPU s)
2: 2017-02-07 20:59:45
2: Benchmark                            Time           CPU Iterations
2: ------------------------------------------------------------------
2: HeapBench/HeapAlloc/32             471 ns        474 ns     137106   2.01184M items/s
2: HeapBench/HeapAlloc/64             457 ns        461 ns     155076   2.06929M items/s
2: HeapBench/HeapAlloc/128            456 ns        459 ns     152351    2.0797M items/s
2: HeapBench/HeapAlloc/256            463 ns        467 ns     145106   2.04163M items/s
2: HeapBench/HeapAlloc/512            441 ns        445 ns     155069   2.14453M items/s
2: HeapBench/HeapAlloc/1024           487 ns        491 ns     151156   1.94119M items/s
2: HeapBench/HeapAlloc/2k             453 ns        456 ns     149706   2.08918M items/s
2: HeapBench/HeapAlloc/4k             455 ns        457 ns     152144   2.08624M items/s
2: HeapBench/HeapAlloc/8k             452 ns        455 ns     144949   2.09461M items/s
2: HeapBench/HeapAlloc/16k            459 ns        462 ns     145667   2.06392M items/s
2: HeapBench/HeapAlloc/32k            495 ns        501 ns     100000   1.90347M items/s
2: HeapBench/HeapAlloc/64k            507 ns        511 ns     100000   1.86489M items/s
2: HeapBench/HeapAlloc/128k           539 ns        542 ns     131798   1.75869M items/s
2: HeapBench/HeapAlloc/256k           616 ns        620 ns     116179   1.53829M items/s
2: HeapBench/HeapAlloc/512k           764 ns        768 ns      94849   1.24218M items/s
2: HeapBench/HeapAlloc/1024k         1069 ns       1071 ns      70113   911.492k items/s
2: HeapBench/HeapDestroy/32           452 ns        455 ns     142537    2.0941M items/s
2: HeapBench/HeapDestroy/64           455 ns        457 ns     150754   2.08634M items/s
2: HeapBench/HeapDestroy/128          455 ns        459 ns     150968   2.07903M items/s
2: HeapBench/HeapDestroy/256          445 ns        447 ns     149368   2.13408M items/s
2: HeapBench/HeapDestroy/512          446 ns        450 ns     145773   2.11985M items/s
2: HeapBench/HeapDestroy/1024         445 ns        448 ns     142651   2.12647M items/s
2: HeapBench/HeapDestroy/2k           458 ns        459 ns     148890   2.07611M items/s
2: HeapBench/HeapDestroy/4k           447 ns        450 ns     150627   2.11897M items/s
2: HeapBench/HeapDestroy/8k           459 ns        462 ns     150773    2.0661M items/s
2: HeapBench/HeapDestroy/16k          538 ns        540 ns     121781   1.76575M items/s
2: HeapBench/HeapDestroy/32k          466 ns        469 ns     145438   2.03181M items/s
2: HeapBench/HeapDestroy/64k          522 ns        524 ns     100000   1.81833M items/s
2: HeapBench/HeapDestroy/128k         541 ns        545 ns     131698   1.74976M items/s
2: HeapBench/HeapDestroy/256k         614 ns        616 ns     116723   1.54787M items/s
2: HeapBench/HeapDestroy/512k         806 ns        811 ns      95129   1.17607M items/s
2: HeapBench/HeapDestroy/1024k       1019 ns       1021 ns      67161   956.391k items/s
```