@sylvanc 2017-02-08
Fantastic. Change looks good, benchmarks look even better!