@sylvanc 2017-02-08
Side note: I am concerned about the CMake dependency, but I left a comment about that in the other PR.