@dipinhora 2017-02-08
Rebased onto updated benchmarks PR #1565.